feat: simplify SSRF protection
