+++ src/config.rs
@@ -7,8 +7,8 @@ use moka::future::Cache;
 use tokio::sync::mpsc::Sender;
 
 use crate::{
+    dns::ValidateDnsResolver,
     endpoints::mods::IndexQueryParams,
-    pin_dns::PinDnsResolver,
     s3_worker::S3WorkerTask,
     storage::{LocalBackend, PrivateDisk, PublicDisk, S3Backend, S3Configuration},
     types::{
@@ -36,7 +36,7 @@ pub struct AppData {
 
     mods_cache: Cache<IndexQueryParams, ApiResponse<PaginatedData<Mod>>>,
     http_client: reqwest::Client,
-    pin_dns_http_client: reqwest::Client,
+    check_dns_http_client: reqwest::Client,
 
     s3_sender: OnceLock<Sender<S3WorkerTask>>,
 }
@@ -86,12 +86,12 @@ pub async fn build_config() -> anyhow::Result<AppData> {
         None
     };
 
-    let pin_dns_http_client = reqwest::Client::builder()
-        .dns_resolver(Arc::new(PinDnsResolver))
+    let check_dns_http_client = reqwest::Client::builder()
+        .dns_resolver(Arc::new(ValidateDnsResolver))
         .pool_max_idle_per_host(4)
         .connect_timeout(Duration::from_secs(10))
-        .read_timeout(Duration::from_secs(30))
-        .redirect(reqwest::redirect::Policy::none())
+        .read_timeout(Duration::from_secs(10))
+        .timeout(Duration::from_secs(30))
         .build()?;
 
     Ok(AppData {
@@ -124,7 +124,7 @@ pub async fn build_config() -> anyhow::Result<AppData> {
             .connect_timeout(Duration::from_secs(10))
             .read_timeout(Duration::from_secs(30))
             .build()?,
-        pin_dns_http_client,
+        check_dns_http_client,
         s3_sender: OnceLock::new(),
     })
 }
@@ -204,19 +204,14 @@ impl AppData {
         &self.http_client
     }
 
-    /// Client that allows pinning DNS queries to a certain ip address.
+    /// Client that validates passed DNS to not contain illegal ip addresses.
     /// Useful for preventing Server Side Request Forgery.
-    ///
-    /// To pin an address, you have to use pin_dns::PINNED_ADDRS.
     ///
     /// Basically, if you have a URL as user input, *always* use this client.
     ///
-    /// The downside is that you get worse DNS performance, doesn't matter when
-    /// security is involved though.
-    ///
     /// For an example, check mod_zip::download()
-    pub fn pin_dns_http_client(&self) -> &reqwest::Client {
-        &self.pin_dns_http_client
+    pub fn check_dns_http_client(&self) -> &reqwest::Client {
+        &self.check_dns_http_client
     }
 
     pub fn init_s3_sender(&self, sender: Sender<S3WorkerTask>) {
