+++ src/main.rs
@@ -13,6 +13,7 @@ mod auth;
 mod cli;
 mod config;
 mod database;
+mod dns;
 mod endpoints;
 mod events;
 mod extractors;
@@ -20,7 +21,6 @@ mod jobs;
 mod logging;
 mod mod_zip;
 mod openapi;
-mod pin_dns;
 mod s3_worker;
 mod storage;
 mod types;
